Add custom data format for attachments (#1)
We currently set the drag format of text/uri-list to allow users to drag
attachments between Front windows. However, this makes it so that
when dragging attachments out of Front to some applications, the
attachment will be interpreted as the URL instead of being inserted as a
file (Word and Outlook for example). To work around this, we use a
custom drag format of x-front-attachment.